+++ Assets/Scripts/CharacterController2D.cs
@@ -56,7 +56,7 @@ public class CharacterController2D : MonoBehaviour
 	}
 
 
-	public void Move(float move, bool crouch, bool jump)
+	public void Move(float move, bool jump)
 	{
 		
 			// Move the character by finding the target velocity
