--- Concepts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fa828f5a40d7e4d4b86b71aea06e2c2f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
