+++ Library/StateCache/SceneView/160bb6-mainStage.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":31,"name":"Sprite Mask","section":"Miscellaneous"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":true,"pivot":{"x":379.85260009765627,"y":212.13140869140626,"z":-467.3207702636719},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":596.93896484375,"orthographic":true}
+﻿{"cameraMode":{"drawMode":31,"name":"Sprite Mask","section":"Miscellaneous"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":true,"pivot":{"x":1.180313229560852,"y":1.587797999382019,"z":-3.70580792427063},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":15.164076805114746,"orthographic":true}
