--- Scenes.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f88b6fb41b45f244aaad9c76fc926884
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
