--- Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a74747987a411a74bb6967e072780006
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
