--- Scripts/CharacterController2D.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d4c625c896c998f45b10f5a4d6da4301
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
