--- Scripts/PlayerController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6d09d6a16fc2cca43ab9de8931fbcae4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
