--- Scripts/StaminaBar.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 95bc6beb61ee0ec4cad5bf09ce0846f8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
