--- Sprites.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f0c82cef72f40e4448e1dec33f7641c2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
